+++ src/game/cards/crops/carrot/carrot.ts
@@ -1,7 +1,9 @@
 import { ICrop } from '../../../types'
 import { baseToCrop } from '../baseToCrop'
 
-export const carrot: ICrop = baseToCrop({
+const carrot: ICrop = baseToCrop({
   id: 'carrot',
   waterToMature: 3,
 })
+
+export default carrot
