--- src/game/cards/crops/carrot/index.ts
@@ -1,1 +1,0 @@
-export * from './carrot'
