+++ src/game/cards/crops/index.ts
@@ -1,2 +1,4 @@
-export * from './carrot'
-export * from './pumpkin'
+import carrot from './carrot/carrot'
+import pumpkin from './pumpkin/pumpkin'
+
+export { carrot, pumpkin }
