--- src/game/cards/crops/pumpkin/index.ts
@@ -1,1 +1,0 @@
-export * from './pumpkin'
