+++ src/game/cards/crops/pumpkin/pumpkin.ts
@@ -1,7 +1,9 @@
 import { ICrop } from '../../../types'
 import { baseToCrop } from '../baseToCrop'
 
-export const pumpkin: ICrop = baseToCrop({
+const pumpkin: ICrop = baseToCrop({
   id: 'pumpkin',
   waterToMature: 4,
 })
+
+export default pumpkin
